+++ src/main/java/com/team_project/shop/web/ApiOrderController.java
@@ -1,45 +1,67 @@
 package com.team_project.shop.web;
 
 import com.team_project.shop.Service.OrderService;
-import com.team_project.shop.Service.UserService_tmp;
+import com.team_project.shop.Service.UsersService;
 import com.team_project.shop.config.auth.LoginUser;
 import com.team_project.shop.config.auth.dto.SessionUser;
+import com.team_project.shop.domain.Response.BaseResponse;
+import com.team_project.shop.domain.Response.CommonResponse;
+import com.team_project.shop.domain.Response.ErrorResponse;
+import com.team_project.shop.domain.Response.StatusCode;
+import com.team_project.shop.domain.order.Orders;
 import com.team_project.shop.network.request.OrdersSaveRequestDto;
 import com.team_project.shop.network.request.OrdersUpdateRequestDto;
 import com.team_project.shop.network.response.OrdersResponseDto;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @RequestMapping("/api/order")
 @RestController
 public class ApiOrderController {
     private final OrderService orderService;
-    private final UserService_tmp userServiceTmp;
+    private final UsersService usersService;
 
     @PostMapping("")
-    public Long save(@LoginUser SessionUser sessionUser, @RequestBody OrdersSaveRequestDto requestDto) {
-        return orderService.save(userServiceTmp.findById(sessionUser.getId()), requestDto);
+    public ResponseEntity<? extends BaseResponse> save(@LoginUser SessionUser sessionUser, @RequestBody OrdersSaveRequestDto requestDto) {
+        Long id = orderService.save(usersService.findById(sessionUser.getId()), requestDto);
+        if(id == 0)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorResponse(StatusCode.BAD_REQUEST.getStatusCode(), "요청이 올바르지 않습니다. 주문을 추가할 수 없습니다."));
+        return ResponseEntity.noContent().build();
     }
 
     @PutMapping("/{orderId}")
-    public Long update(@LoginUser SessionUser sessionUser, @RequestBody OrdersUpdateRequestDto requestDto, @PathVariable Long orderId) {
-        if(!orderService.findById(orderId).equals(sessionUser.getId()))
-            return -1l;
-        return orderService.update(orderId, requestDto);
+    public ResponseEntity<? extends BaseResponse> update(@LoginUser SessionUser sessionUser, @RequestBody OrdersUpdateRequestDto requestDto, @PathVariable Long orderId) {
+        Optional<OrdersResponseDto> order = orderService.findById(orderId);
+        if(!order.isPresent())
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse(StatusCode.NOT_FOUND.getStatusCode(), "존재하지 않는 주문입니다."));
+        if(!order.get().getUser().getId().equals(sessionUser.getId()))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse(StatusCode.FORBIDDEN.getStatusCode(), "해당 주문에 접근할 수 없습니다."));
+        orderService.update(orderId, requestDto);
+        return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{orderId}")
-    public Long delete(@LoginUser SessionUser sessionUser, @PathVariable Long orderId) {
-        //이거 서비스에서 수정하기!!! 배송완료 같은거면 취소불가하는 로직 짜서 리턴값 달리 주기
-        return orderService.delete(orderId);
+    public ResponseEntity<? extends BaseResponse> delete(@LoginUser SessionUser sessionUser, @PathVariable Long orderId) {
+        Optional<OrdersResponseDto> order = orderService.findById(orderId);
+        if(!order.isPresent())
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse(StatusCode.NOT_FOUND.getStatusCode(), "존재하지 않는 주문입니다."));
+        if(!order.get().getUser().getId().equals(sessionUser.getId()))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse(StatusCode.FORBIDDEN.getStatusCode(), "해당 주문에 접근할 수 없습니다."));
+        Long id = orderService.delete(orderId);
+        if(id == 0)
+            return ResponseEntity.ok().body(new CommonResponse(null, "주문을 취소할 수 없습니다."));
+        return ResponseEntity.noContent().build();
     }
 
 
     @GetMapping("/my")
-    public List<OrdersResponseDto> findByUserId(@LoginUser SessionUser sessionUser) {
-        return orderService.findByUserId(sessionUser.getId());
+    public ResponseEntity<? extends BaseResponse> findByUser(@LoginUser SessionUser sessionUser) {
+        return ResponseEntity.ok().body(new CommonResponse(orderService.findByUserId(sessionUser.getId()),"주문목록 조회 성공"));
     }
 }
